Add minimal integration tests (#3)
The pipeline currently fails as there's no integration test defined.

We'll add a minimal test that deploys the vsphere chart. The actual
k8s deployment is expcted to enter a crash loop as it expects
an actual vsphere environment, for which reason the assertions
are commented out for the time being.